Add missing counter example to `synthetic/sum_is_commutative_klee_float_bug.x86_64`.
Unfortunately there appears to be a bug in Clang 3.4 where the line
location in the debug info is on the wrong line when the loop head is
spread across multiple lines. So this patch reworks the code to avoid
this.

The bug appears to be fixed in newer Clang versions (the location was
right in Clang 3.9) so there is no need to report this issue upstream.

This fixes # 4.