Bug fix for sorted_search benchmark.
The benchmark would fail to terminate (e.g. if `lo=3`, `hi=4`)
in certain conditions.

It was unclear to me if the range was supposed to be

- Closed [lo,hi]
- Half open [lo, hi)

Part of the code was written as if the range under consideration was
half open but the mid point calculation was written as if the range
was closed.

So I have switched to a closed range and have fixed the assignments
to `mid`.